Prototype/BirdInfoGen/requirements.txt
flask
Flask-Cors
numpy
pandas
faiss-cpu
sentence-transformers
scikit-learn
rapidfuzz
transformers
torch
